requirements.txt
tqdm
python-dotenv
openai
pyyaml
